Standardized variable formats and sorting.
Added options:
	--var-name <name>       variable name
	--freqs <freq_format>	see RF::Component->at() and PDL::IO::Touchstone::m_interpolate()

For these see below:
	--sort	<var_format>
	--mdif-var-format <mixed_format>

The process of this refactor made --var-code (or --var-literal) optional, but
it is recommended if you wish to use 'value' or 'value_unit' in your output
formatting:

The MDIF format supports variable definitions for selecting a component in your
EDA software.  At the moment only a single variable will be omitted, for
example:

	VAR pF=100;

To accomplish this would would write a --mdif-var-format as follows:

	--var-name pF --mdif-var-format '{capacitance}'

The syntax for the variable format is as follows:

	{[format-specifier:]function[operator]}

* format-specifier: If a printf format specifier is used then it will be passed
  through `sprintf`

* function may be one of:
   1. any function provided by RF::Component
   2. or "idx" which will be incremented when writing to the MDIF file.

Note that if `function` returns a PDL with more than one value then it will
be averaged using an arithmetic mean.  If you want a specific index then see
the [N] operator.

* operator may be one of:
	1. [N]: The Nth index into a PDL slice.  Negative numbers count from the end.
	2. /N : divide by N
	3. *N : multiply by N.
  where N can be any floating point value

Example 1:
	{idx}. {value}{value_unit} X={%.2f:reactance[0]} {model} SRF={%.2f:srf/1e9}GHz

	VAR pF="0. 1.1pF X=-2893.38 GRM1555C1H1R1WA01 SRF=7.74GHz
	VAR pF="1. 1.2pF X=-2652.23 GRM1555C1H1R2WA01 SRF=7.51GHz
	VAR pF="2. 1.3pF X=-2448.19 GRM1555C1H1R3WA01 SRF=7.28GHz

Example 2:
	{idx}. {value}{value_unit} {model} ({%.2f:freqs[0]/1e6}-{%.2f:freqs[-1]/1e6} MHz)

	VAR pF="0. 1pF GRM1555C1H1R0WA01 (100.00-10050.00 MHz)"
	VAR pF="1. 1.1pF GRM1555C1H1R1WA01 (100.00-4300.00 MHz)"
	VAR pF="2. 1.2pF GRM1555C1H1R2WA01 (100.00-4300.00 MHz)"

Sorting:

You can specify any variable format in --sort, just omit {}. For example:
	--sort value
or
	--sort 'reactance[0]'

Currently the default to sort by 'reactance[0]'.