Only call get_stock if the part has not been seen.
It was making unnecessary of API calls.  the the mdf script attempts to
minimize API  calls when --unique as specified  if a parts matching the sort
field has already been found. It was making a call to get_stock for all parts,
now it only makes that call if the part has not been seen.

Refactored octopart-specific options into its own hash
Added octopart options:
	--octopart-debug
	--octopart-query-limit